Avoid using log.* functions for error printing
In the `main` function using the `log` package prints a lot of metadata
around the error/message that it is not very useful considering that the
program is invoked directly.

For the normal assuan protocol invokations inside our client the logger
is still used, because the metadata is useful and also it gets
automatically written into `/tmp/pinentry-touchid.log`.