Additional checks when running with `--check`
Extend the checks that are run when the `--check` flag is provided.
Previously we only checked that the binary existed in the current
`$PATH`.

If the binary is a symlink it will be resolved and the end file will
be compared with `pinentry-mac`.

This should prevent the issue reported in #3 where default gpg
installation via homebrew will break because the output of `gpgconf`
contains:

```
pinentry:Passphrase Entry:/usr/local/opt/pinentry/bin/pinentry
```

At the same time `/usr/local/opt/pinentry/bin/pinentry` by default
points to `pinentry-curses` which means that pinentry-touchid is unable
to call the fallback pinentry program entirely.