Store the log file in $TMPDIR instead of /tmp
Instead of using `/tmp` as the base we now use the value of `$TMPDIR`
which is user-scoped. This means that the path of the log file is no
longer `/tmp/pinentry-touchid.log` but `$TMPDIR/pinentry-touchid.log`.

This is particularly important in multiuser environments where the first
user would get ownership of the file and actions from other users would
fail.

Fixes #13