Add -fix flag for automatically fixing the pinentry symlink
When installing gnupg, homebrew automatically configures the path of the
pinentry binary to the `opt_bin` path of the formula instead of the
symlink in `/usr/local/bin`
https://github.com/Homebrew/homebrew-core/blob/08d0d1cc331d570a31d5ef89e695ad283c75d4f6/Formula/gnupg.rb#L54
this means that actual path returned by `gpgconf` ends up
looking like: /usr/local/opt/pinentry/bin/pinentry.

Fixing the symlink in `/usr/local/bin` doesn't always solve the issue
because `gpgconf` is still reporting the `/usr/local/opt` path.

The new flag changes automatically this symlink to point to
pinentry-mac.

I tried unsuccessfully to include this as part of the automatic Formula
installation but it is not possible because during installation the
formula cannot modify any external path.