Avoid sending trailing spaces in io.WriteLine
The previous implementation was always sending a trailing space when
no params were provided. For instance when the `OK` command gets sent
back.

Although the gpg-agent seems to be able to deal with this some other
implementations like the pinentry crate had a problem parsing the
response. Some apps like `rage` use this crate.